src/requirements.txt
PyYAML==6.0
termcolor==1.1.0
cryptography==36.0.1
redis==4.1.4
websocket-client==1.3.1
orjson==3.6.7
pandas-market-calendars==3.4
git+ssh://git@github.com/stopdesign/ibkr_web_api@develop#egg=ibkr_web_api
click==8.0.4
